Move Parser into internal package
Closes #2549
