+++ testng-core/src/test/java/test/methodinterceptors/MethodInterceptorTest.java
@@ -12,7 +12,7 @@ import org.testng.ITestResult;
 import org.testng.TestListenerAdapter;
 import org.testng.TestNG;
 import org.testng.annotations.Test;
-import org.testng.xml.Parser;
+import org.testng.xml.internal.Parser;
 import org.testng.xml.XmlSuite;
 import org.xml.sax.SAXException;
 import test.SimpleBaseTest;
