+++ testng-core/src/test/java/test/suites/github1533/Github1533Test.java
@@ -9,7 +9,7 @@ import org.testng.ITestNGListener;
 import org.testng.TestNG;
 import org.testng.TestNGException;
 import org.testng.annotations.Test;
-import org.testng.xml.Parser;
+import org.testng.xml.internal.Parser;
 import org.testng.xml.XmlSuite;
 import test.SimpleBaseTest;
