+++ testng-core/src/test/java/test/xml/issue2231/IssueTest.java
@@ -10,7 +10,7 @@ import java.util.List;
 import java.util.Map;
 import org.testng.TestListenerAdapter;
 import org.testng.annotations.Test;
-import org.testng.xml.Parser;
+import org.testng.xml.internal.Parser;
 import org.testng.xml.XmlClass;
 import org.testng.xml.XmlSuite;
 import org.testng.xml.XmlSuite.ParallelMode;
